pyproject.toml
[project]
name = "rag-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "bs4>=0.0.2",
    "langchain>=1.0.2",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=3.0.0",
    "langchain-qdrant>=1.1.0",
    "langchain-text-splitters>=1.0.0",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.15.1",
]
